+++ components/bluetooth_proxy/src/lib.rs
@@ -23,8 +23,9 @@ pub struct BluetoothProxyConfig {
 }
 
 config_template!(
-    mdns,
+    bluetooth_proxy,
     Option<BluetoothProxyConfig>,
+    NoConfig,
     NoConfig,
     NoConfig,
     NoConfig,
@@ -57,7 +58,10 @@ impl Module for UbiHomePlatform {
     ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn std::error::Error>>> + Send + 'static>>
     {
         let config = self.config.clone();
-        info!("Starting Bluetooth Proxy with config: {:?}", config.mdns);
+        info!(
+            "Starting Bluetooth Proxy with config: {:?}",
+            config.bluetooth_proxy
+        );
         Box::pin(async move {
             // TODO: Check if bluetooth is enabled