+++ components/core/src/lib.rs
@@ -4,6 +4,7 @@ pub mod features;
 pub mod internal;
 pub mod light;
 pub mod mapper;
+pub mod number;
 pub mod sensor;
 pub mod switch;
 pub mod utils;
@@ -89,6 +90,14 @@ pub enum ChangedMessage {
         green: Option<f32>,
         blue: Option<f32>,
     },
+    NumberValueChange {
+        key: String,
+        value: f32,
+    },
+    NumberValueCommand {
+        key: String,
+        value: f32,
+    },
     BluetoothProxyMessage(BluetoothProxyMessage),
 }
 
@@ -132,6 +141,14 @@ pub enum PublishedMessage {
         green: Option<f32>,
         blue: Option<f32>,
     },
+    NumberValueChanged {
+        key: String,
+        value: f32,
+    },
+    NumberValueCommand {
+        key: String,
+        value: f32,
+    },
     BluetoothProxyMessage(BluetoothProxyMessage),
 }
 
@@ -170,7 +187,8 @@ macro_rules! config_template {
         $binary_sensor_extension:ident,
         $sensor_extension:ident,
         $switch_extension:ident,
-        $light_extension:ident) => {
+        $light_extension:ident,
+        $number_extension:ident) => {
         use duration_str::deserialize_option_duration;
         use garde::Validate;
         use ubihome_core::UbiHome;
@@ -181,6 +199,7 @@ macro_rules! config_template {
         use ubihome_core::template_mapper_kind;
         use ubihome_core::template_mapper_new;
         use ubihome_core::template_mapper3;
+        use ubihome_core::template_number;
         use ubihome_core::template_switch;
 
         // template_button!($component_name, $button_extension);
@@ -188,6 +207,7 @@ macro_rules! config_template {
         template_entity!($component_name, $sensor_extension);
         template_switch!($component_name, $switch_extension);
         template_light!($component_name, $light_extension);
+        template_number!($component_name, $number_extension);
 
         // template_mapper!(map_sensor, Sensor);
         template_mapper_kind!(map_sensor, $component_name, Sensor);
@@ -195,6 +215,7 @@ macro_rules! config_template {
         template_mapper_new!(map_binary_sensor, $component_name, $binary_sensor_extension);
         template_mapper!(map_switch, Switch);
         template_mapper!(map_light, Light);
+        template_mapper!(map_number, Number);
 
         #[derive(Clone, Deserialize, Debug, Validate)]
         #[garde(allow_unvalidated)]
@@ -224,6 +245,9 @@ macro_rules! config_template {
             #[serde(default, deserialize_with = "map_light")]
             #[garde(dive)]
             pub light: Option<HashMap<String, Light>>,
+
+            #[serde(default, deserialize_with = "map_number")]
+            pub number: Option<HashMap<String, Number>>,
         }
     };
 }