+++ components/evdev/src/lib.rs
@@ -23,6 +23,7 @@ config_template!(
     NoConfig,
     NoConfig,
     NoConfig,
+    NoConfig,
     NoConfig
 );