+++ components/gpio/src/lib.rs
@@ -50,6 +50,7 @@ config_template!(
     GpioBinarySensorConfig,
     NoConfig,
     NoConfig,
+    NoConfig,
     NoConfig
 );