+++ components/illuminance/src/lib.rs
@@ -45,6 +45,7 @@ config_template!(
     NoConfig,
     AmbientLightSensorConfig,
     NoConfig,
+    NoConfig,
     NoConfig
 );