+++ components/power_utils/src/lib.rs
@@ -75,6 +75,7 @@ config_template!(
     NoConfig,
     NoConfig,
     NoConfig,
+    NoConfig,
     NoConfig
 );