+++ components/web_server/src/lib.rs
@@ -48,6 +48,7 @@ config_template!(
     NoConfig,
     NoConfig,
     NoConfig,
+    NoConfig,
     NoConfig
 );