+++ tests/conftest.py
@@ -1,6 +1,7 @@
-from typing import Any, Generator
-import pytest
+from collections.abc import Generator
+from typing import Any
 
+import pytest
 from mock_file import IOMock, IOMockFactory
 
 
@@ -13,6 +14,6 @@ def io_mock() -> Generator[IOMock, Any, Any]:
 
 @pytest.fixture(scope="function")
 def io_mock_factory() -> Generator[IOMockFactory, Any, Any]:
-    mock_file = IOMockFactory(base_path="")
-    yield mock_file
-    mock_file.cleanup()
+    mock_factory = IOMockFactory(base_path="")
+    yield mock_factory
+    mock_factory.cleanup()