+++ tests/platforms/mqtt/mqtt_device_info_test.py
@@ -1,9 +1,11 @@
-from asyncio import sleep
 import json
+from asyncio import sleep
 from unittest.mock import Mock
-from utils import UbiHome
-from paho.mqtt.client import Client
+
 import pytest
+from paho.mqtt.client import Client
+from utils import UbiHome
+
 
 @pytest.mark.timeout(60)
 async def test_run(mqtt_client: Client, mqtt_connection):
@@ -13,14 +15,14 @@ ubihome:
   name: {name}
 
 mqtt:
-  broker: {mqtt_connection['host']}
-  port: {mqtt_connection['port']}
+  broker: {mqtt_connection["host"]}
+  port: {mqtt_connection["port"]}
 """
 
     mqtt_client.subscribe("#")
     mock = Mock()
     mqtt_client._on_message = mock
-    async with UbiHome("run", config=DEVICE_INFO_CONFIG) as ubihome:
+    async with UbiHome("run", config=DEVICE_INFO_CONFIG):
         await sleep(1)
         mock.assert_called_once()
         message = mock.call_args.args[2]