+++ components/illuminance/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [0.13.0](https://github.com/UbiHome/UbiHome/compare/v0.12.0...ubihome-illuminance-v0.13.0) (2026-05-10)
+
+
+### Features
+
+* add `number` platform support for shell component ([#112](https://github.com/UbiHome/UbiHome/issues/112)) ([1799769](https://github.com/UbiHome/UbiHome/commit/1799769ccddf70cfff0f73e764e659cdef5280cd))
+
+## [0.12.0](https://github.com/UbiHome/UbiHome/compare/v0.11.1...ubihome-illuminance-v0.12.0) (2026-04-16)
+
+
+### Features
+
+* support `accuracy_decimals` in sensor entity configuration ([#113](https://github.com/UbiHome/UbiHome/issues/113)) ([9ccbcd4](https://github.com/UbiHome/UbiHome/commit/9ccbcd4efd39304b4a1960d548eb33b6e282ff41))
+
+## [0.11.1](https://github.com/UbiHome/UbiHome/compare/v0.11.0...ubihome-illuminance-v0.11.1) (2026-04-14)
+
+## [0.11.0](https://github.com/UbiHome/UbiHome/compare/ubihome-illuminance-v0.10.4...ubihome-illuminance-v0.11.0) (2026-04-14)
+
+
+### Features
+
+* add illuminance sensor ([#83](https://github.com/UbiHome/UbiHome/issues/83)) ([fc24c29](https://github.com/UbiHome/UbiHome/commit/fc24c2945849140969b0d7cf6a308fd91852d027))
+
+## [0.10.4](https://github.com/UbiHome/UbiHome/compare/v0.10.0...ubihome-illuminance-v0.10.4) (2026-04-14)
+
 ## [0.10.0](https://github.com/UbiHome/UbiHome/compare/v0.9.0...ubihome-illuminance-v0.10.0) (2025-12-29)